+++ src/main/java/subway/member/ui/MemberController.java
@@ -3,7 +3,7 @@ package subway.member.ui;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import subway.auth.domain.AuthenticationPrincipal;
-import subway.member.application.MemberService;
+import subway.member.service.MemberService;
 import subway.member.domain.LoginMember;
 import subway.member.dto.MemberRequest;
 import subway.member.dto.MemberResponse;
